♻️ F13: don't mix component and function in exports of a file for vite to hot-reload properly
