+++ src/context/useOrderContext.ts
@@ -1,5 +1,33 @@
-import { useContext } from "react";
-import { OrderContext } from "./OrderContext";
+import { createContext, Dispatch, SetStateAction, useContext } from "react";
+import { TabProps } from "src/components/pages/order/Admin/AdminTabs/AdminTabs";
+import { MenuItem } from "src/fakeData/fakeMenu";
+import { BasketItem } from "../hooks/useBasket";
+
+type OrderContextType = {
+  isAdminMode: boolean;
+  setIsAdminMode: Dispatch<SetStateAction<boolean>>;
+  isPanelOpen: boolean;
+  setIsPanelOpen: Dispatch<SetStateAction<boolean>>;
+  tabs: TabProps[];
+  setTabs: Dispatch<SetStateAction<TabProps[]>>;
+  selectedTab: () => TabProps | undefined;
+  selectTab: (index: string) => void;
+  menu: MenuItem[];
+  addItemToMenu: (item: Partial<MenuItem>) => void;
+  removeItemFromMenu: (id: number) => void;
+  loadMenu: (userId: string) => void;
+  resetMenu: () => void;
+  selectedItem: MenuItem;
+  setSelectedItemById: (id: number) => void;
+  unSelectItem: () => void;
+  updateItem: (item: MenuItem) => void;
+  basket: BasketItem[];
+  addItemToBasket: (id: number) => void;
+  removeItemFromBasket: (id: number) => void;
+  refInputName: React.RefObject<HTMLInputElement>;
+};
+
+export const OrderContext = createContext<OrderContextType | undefined>(undefined);
 
 // custom hook to use the order context in any component with only one import
 export const useOrderContext = () => {
